requirements_dev.txt
bump-my-version==0.6.0
wheel==0.36.2
watchdog==2.0.2
ruff>=0.3.3
black==22.3.0
tox==3.23.0
coverage==5.5
Sphinx==3.5.4
twine==1.11.0

pytest==8.4.0
pytest-asyncio==0.14.0
pytest-cov==2.11.1
pytest-mock==3.5.1
pluggy>=1.5.0
python-coveralls==2.9.3
requests==2.25.1
parsel==1.6.0
sphinx-rtd-theme==0.5.2
testfixtures==6.17.1
pytest-aiohttp==0.3.0
docutils==0.12
factory-boy==3.2.0
aioresponses==0.7.6
motor~=3.4.0
